UPDATE airtv-serve
Currently trying to get all of the basis elements in, but it's going quite smoothly.

All I need to get fixed is the build on webhook method.

And then I just need to add the api layer, but that will be relatively easy, the hard part is just getting all of these seperate layers working together.